# backend/app/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  types.BriefUser:
    description: 简要的用户信息展示
    properties:
      avatar_url:
        example: https://127.0.0.1:8080/api/v1/user/auth/avatar/1
        type: string
      backimage_url:
        example: https://127.0.0.1:8080/api/v1/user/auth/backimage/1
        type: string
      campus:
        example: 翡翠湖校区
        type: string
      create_at:
        example: "2025-04-27T16:10:08.5Z"
        type: string
      department:
        example: 数学学院
        type: string
      grade:
        example: "2022"
        type: string
      id:
        example: 1
        type: integer
      nickname:
        example: 傅益忠
        type: string
      sex:
        example: 男
        type: string
      signature:
        example: 这人啥也没说
        type: string
      tags:
        description: 注意：原结构体类型应为[]string
        example:
        - 学习
        - 运动
        - 音乐
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  title: 接口文档
  version: "1.0"
paths:
...
  /user/auth/{id}:
    get:
      description: 根据用户ID返回用户
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/types.BriefUser'
      security:
      - BearerAuth: []
      summary: 获取用户数据
      tags:
      - User模块
...
securityDefinitions:
  BearerAuth:
    description: Bearer Token 格式：`Bearer {token}`（注意空格）
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"